Преобразует функциональные компоненты в стейт-компоненты
