+++ src/view/event-list/events-list.js
@@ -1,6 +1,25 @@
-export const createEventsList = (dataLength) => {
+import {createElement} from "../../utils/render";
+import {createEventsList} from "./templates/create-events-list";
 
-  return (
-    dataLength && `<ul class="trip-events__list"></ul>` || ``
-  );
-};
+export default class EventsList {
+  constructor(dataLength) {
+    this._dataLength = dataLength;
+    this._element = null;
+  }
+
+  getTemplate() {
+    return createEventsList(this._dataLength);
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
+
+    return this._element;
+  }
+
+  removeElement() {
+    this._element = null;
+  }
+}
