+++ src/view/event-list/templates/create-events-list.js
@@ -1,0 +1,6 @@
+export const createEventsList = (dataLength) => {
+
+  return (
+    dataLength && `<ul class="trip-events__list"></ul>` || ``
+  );
+};
