+++ src/view/warning/templates/create-warning-template.js
@@ -1,0 +1,6 @@
+export const createWarningTemplate = (warning) => {
+
+  return (
+    `<p class="trip-events__msg">${warning}</p>`
+  );
+};
