+++ src/view/warning/warning.js
@@ -1,6 +1,25 @@
-export const createWarningTemplate = (warning) => {
+import {createElement} from "../../utils/render";
+import {createWarningTemplate} from "./templates/create-warning-template";
 
-  return (
-    `<p class="trip-events__msg">${warning}</p>`
-  );
-};
+export default class Warning {
+  constructor(warning) {
+    this._warning = warning;
+    this._element = null;
+  }
+
+  getTemplate() {
+    return createWarningTemplate(this._warning);
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
+
+    return this._element;
+  }
+
+  removeElement() {
+    this._element = null;
+  }
+}
